# [Backport 6.1] replica/table: check memtable before discarding tombstone during read