@scylladb-promoter 2024-10-23
#### :red_circle: CI State: FAILURE

:x: - [Build](https://jenkins.scylladb.com/job/scylla-6.1/job/scylla-ci/136/)
#### Build Details:
- Duration: 31 min
- Builder: i-091dfdf048620c882 (m5ad.8xlarge)