@scylladb-promoter 2024-10-24
#### :green_circle: CI State: SUCCESS

:white_check_mark: - [Build](https://jenkins.scylladb.com/job/scylla-6.1/job/scylla-ci/137/)
:white_check_mark: - [Unit Tests Custom](https://jenkins.scylladb.com/job/scylla-6.1/job/scylla-ci/137/)
**The following new/updated tests ran 100 times for each mode:**
:small_blue_diamond: topology_custom/test_compacting_reader_tombstone_gc
:white_check_mark: - [Container Test](https://jenkins.scylladb.com/job/scylla-6.1/job/scylla-ci/137/)
:white_check_mark: - [dtest with topology changes](https://jenkins.scylladb.com/job/scylla-6.1/job/gating-dtest-release-with-consistent-topology-changes/267/)
:white_check_mark: - [dtest](https://jenkins.scylladb.com/job/scylla-6.1/job/gating-dtest-release/274/)
:white_check_mark: - [Unit Tests](https://jenkins.scylladb.com/job/scylla-6.1/job/scylla-ci/137/)
#### Build Details:
- Duration: 5 hr 53 min
- Builder: i-08fe4c1d2fc948c2c (m5ad.8xlarge)